requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pytest==8.3.3
httpx==0.27.2
